backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
google-genai
python-multipart
deepgram-sdk
python-dotenv
